Refactored backend to follow Flask factory/blueprint pattern
