+++ backend/requirements.txt
@@ -2,3 +2,4 @@ Flask==2.0.1
 flask_cors==4.0.0
 werkzeug==2.0.1
 gunicorn==21.2.0
+boto3==1.34.6
