+++ backend/settings_module.py
@@ -1,0 +1,6 @@
+# settings_module.py
+
+class ConfigClass:
+    DEBUG = False  # Enable or disable Flask's debug mode
+    HOST = '0.0.0.0'
+    PORT = 8000
